Cleanup stats. Make more definitive add errors
