+++ collector/interfaces.go
@@ -1,11 +1,5 @@
 package collector
 
-import (
-	"fmt"
-
-	"go.aporeto.io/trireme-lib/policy"
-)
-
 // Flow event description
 const (
 	// FlowReject indicates that a flow was rejected
@@ -32,26 +26,6 @@ const (
 	APIPolicyDrop = "api"
 )
 
-// Container event description
-const (
-	// ContainerStart indicates a container start event
-	ContainerStart = "start"
-	// ContainerStop indicates a container stop event
-	ContainerStop = "stop"
-	// ContainerCreate indicates a container create event
-	ContainerCreate = "create"
-	// ContainerDelete indicates a container delete event
-	ContainerDelete = "delete"
-	// ContainerUpdate indicates a container policy update event
-	ContainerUpdate = "update"
-	// ContainerFailed indicates an event that a container was stopped because of policy issues
-	ContainerFailed = "forcestop"
-	// ContainerIgnored indicates that the container will be ignored by Trireme
-	ContainerIgnored = "ignore"
-	// ContainerDeleteUnknown indicates that policy for an unknown  container was deleted
-	ContainerDeleteUnknown = "unknowncontainer"
-)
-
 const (
 	// PolicyValid Normal flow accept
 	PolicyValid = "V"
@@ -72,87 +46,4 @@ type EventCollector interface {
 
 	// CollectUserEvent  collects a user event
 	CollectUserEvent(record *UserRecord)
-}
-
-// EndPointType is the type of an endpoint (PU or an external IP address )
-type EndPointType byte
-
-const (
-	// EndPointTypeExternalIP indicates that the endpoint is an external IP address
-	EndPointTypeExternalIP EndPointType = iota
-	// EnpointTypePU indicates that the endpoint is a PU.
-	EnpointTypePU
-	// EndpointTypeClaims indicates that the endpoint is of type claims.
-	EndpointTypeClaims
-)
-
-func (e *EndPointType) String() string {
-
-	switch *e {
-	case EndPointTypeExternalIP:
-		return "ext"
-	case EnpointTypePU:
-		return "pu"
-	case EndpointTypeClaims:
-		return "claims"
-	}
-
-	return "pu" // backward compatibility (CS: 04/24/2018)
-}
-
-// EndPoint is a structure that holds all the endpoint information
-type EndPoint struct {
-	ID         string
-	IP         string
-	URI        string
-	HTTPMethod string
-	UserID     string
-	Type       EndPointType
-	Port       uint16
-}
-
-// FlowRecord describes a flow record for statistis
-type FlowRecord struct {
-	ContextID        string
-	Source           *EndPoint
-	Destination      *EndPoint
-	Tags             *policy.TagStore
-	DropReason       string
-	PolicyID         string
-	ObservedPolicyID string
-	ServiceType      policy.ServiceType
-	ServiceID        string
-	Count            int
-	Action           policy.ActionType
-	ObservedAction   policy.ActionType
-	L4Protocol       uint8
-}
-
-func (f *FlowRecord) String() string {
-	return fmt.Sprintf("<flowrecord contextID:%s count:%d sourceID:%s destinationID:%s sourceIP: %s destinationIP:%s destinationPort:%d action:%s mode:%s>",
-		f.ContextID,
-		f.Count,
-		f.Source.ID,
-		f.Destination.ID,
-		f.Source.IP,
-		f.Destination.IP,
-		f.Destination.Port,
-		f.Action.String(),
-		f.DropReason,
-	)
-}
-
-// ContainerRecord is a statistics record for a container
-type ContainerRecord struct {
-	ContextID string
-	IPAddress policy.ExtendedMap
-	Tags      *policy.TagStore
-	Event     string
-}
-
-// UserRecord reports a new user access. These will be reported
-// periodically.
-type UserRecord struct {
-	ID     string
-	Claims []string
 }
